Swarm should handle devices without asic endpoint (#1089)
* Swarm should handle devices without asic endpoint

This is best effort backwards compatible leniency towards devices that don't have the /api/system/asic endpoint yet. It also points to a way to move towards a more organised api. In this version asicCount is moved from info to asic endpoint.

* Clean up asicCount on info endpoint, fix openapi.yaml

* Make the call to `asic` optional.

Not using SystemModule calls to fetch info, as other versions might not adhere to the same structure.

Fallback from familyName to deviceModel to 'Other'.

* Change familyName to deviceModel

* Revert unrelated rename

* Whitespace

* Remove comment