Isolate NVS usage into task (#1286)
* NVS writer task

Delegate all NVS writes to a separate task. This solves the conflict of task heap allocations in Spiram with NVS usage.

To achieve this, the Settings structure that was present in http_server is extracted, and extended to hold all NVS fields.

All default values are also defined in the settings, and used if NVS is not available or the field is not set yet. This has as extra benefit that the getters do not need a default_value anymore, which in many cases, was duplicated several times.

Finally added a nvs_get/set_bool, for ease of use in the code.

* Some bugs

* Reorder enum

* Fix custom checks in http_server

* Proper casting of u64 and true check on bool

* Move task heap of non-mining tasks to spiram

Remove nvs_config logging

* Cleanups

* Add task heap

* More cleanups

* Sleep at end of self test

* Fix nvs_open failure handling